Refactor useSeqGenerator away from custom hook
